feat: Simplify endpoints
