+++ src/main/kotlin/uk/gov/gdx/datashare/resource/AdminController.kt
@@ -1,0 +1,49 @@
+package uk.gov.gdx.datashare.resource
+
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import kotlinx.coroutines.flow.toList
+import org.springframework.http.MediaType
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.validation.annotation.Validated
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+import uk.gov.gdx.datashare.repository.EgressEventDataRepository
+import uk.gov.gdx.datashare.repository.IngressEventDataRepository
+
+@RestController
+@RequestMapping("/admin", produces = [MediaType.APPLICATION_JSON_VALUE])
+//TODO-https://trello.com/c/VgQeakpA: Add admin controll
+@PreAuthorize("hasAnyAuthority('SCOPE_pubsub/maintain')")
+@Validated
+class AdminController(
+  private val egressEventDataRepository: EgressEventDataRepository,
+  private val ingressEventDataRepository: IngressEventDataRepository
+) {
+  @GetMapping("/events/egress")
+  @Operation(
+    summary = "Get Egress Events",
+    description = "Need scope of pubsub/maintain",
+    responses = [
+      ApiResponse(
+        responseCode = "200",
+        description = "Egress Event"
+      )
+    ]
+  )
+  suspend fun getEgressEvents() = egressEventDataRepository.findAll().toList()
+
+  @GetMapping("/events/ingress")
+  @Operation(
+    summary = "Get Ingress Events",
+    description = "Need scope of pubsub/maintain",
+    responses = [
+      ApiResponse(
+        responseCode = "200",
+        description = "Ingress Event"
+      )
+    ]
+  )
+  suspend fun getIngressEvents() = ingressEventDataRepository.findAll().toList()
+}
