+++ src/main/kotlin/uk/gov/gdx/datashare/resource/EventsController.kt
@@ -1,5 +1,6 @@
 package uk.gov.gdx.datashare.resource
 
+import com.fasterxml.jackson.annotation.JsonInclude
 import io.micrometer.core.instrument.Counter
 import io.micrometer.core.instrument.MeterRegistry
 import io.swagger.v3.oas.annotations.Operation
@@ -12,6 +13,7 @@ import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
+import uk.gov.gdx.datashare.service.DataReceiverService
 import uk.gov.gdx.datashare.service.EventDataService
 import uk.gov.gdx.datashare.service.EventNotification
 import java.time.LocalDateTime
@@ -23,28 +25,11 @@ import java.util.*
 @Validated
 class EventsController(
   private val eventDataService: EventDataService,
+  private val dataReceiverService: DataReceiverService,
   meterRegistry: MeterRegistry,
 ) {
   private val callsToPollCounter: Counter = meterRegistry.counter("API_CALLS.CallsToPoll")
-
-  @DeleteMapping("/{id}")
-  @Operation(
-    summary = "Event Delete API - Delete event data",
-    description = "The event ID is the UUID received off the queue, Need scope of data_retriever/read",
-    responses = [
-      ApiResponse(
-        responseCode = "204",
-        description = "Event deleted"
-      )
-    ]
-  )
-  suspend fun deleteEvent(
-    @Schema(description = "Event ID", required = true)
-    @PathVariable id: UUID,
-  ): ResponseEntity<Void> {
-    eventDataService.deleteEvent(id)
-    return ResponseEntity<Void>(HttpStatus.NO_CONTENT)
-  }
+  private val ingestedEventsCounter: Counter = meterRegistry.counter("API_CALLS.IngestedEvents")
 
   @GetMapping
   @Operation(
@@ -82,4 +67,61 @@ class EventsController(
     callsToPollCounter.increment()
     eventDataService.getEvents(eventTypes, startTime, endTime)
   }
+
+  @PreAuthorize("hasAnyAuthority('SCOPE_data_receiver/notify')")
+  @PostMapping
+  @Operation(
+    summary = "Send ingress events to GDS - The 'Source' of the event - this could be HMPO or DWP for example",
+    description = "Scope is data_receiver/notify",
+    responses = [
+      ApiResponse(
+        responseCode = "201",
+        description = "Data Accepted"
+      )
+    ]
+  )
+  suspend fun publishEvent(
+    @Schema(
+      description = "Event Payload",
+      required = true,
+      implementation = EventToPublish::class,
+    )
+    @RequestBody eventPayload: EventToPublish,
+  ) = run {
+    dataReceiverService.sendToDataProcessor(eventPayload)
+    ingestedEventsCounter.increment()
+  }
+
+  @DeleteMapping("/{id}")
+  @Operation(
+    summary = "Event Delete API - Delete event data",
+    description = "The event ID is the UUID received off the queue, Need scope of data_retriever/read",
+    responses = [
+      ApiResponse(
+        responseCode = "204",
+        description = "Event deleted"
+      )
+    ]
+  )
+  suspend fun deleteEvent(
+    @Schema(description = "Event ID", required = true)
+    @PathVariable id: UUID,
+  ): ResponseEntity<Void> {
+    eventDataService.deleteEvent(id)
+    return ResponseEntity<Void>(HttpStatus.NO_CONTENT)
+  }
 }
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@Schema(description = "Event Payload for GDX")
+data class EventToPublish(
+  @Schema(description = "Type of event", required = true, example = "DEATH_NOTIFICATION", allowableValues = [ "DEATH_NOTIFICATION", "LIFE_EVENT"])
+  val eventType: String,
+  @Schema(description = "Date and time when the event took place, default is now", required = false, type = "date-time", example = "2021-12-31T12:34:56")
+  @DateTimeFormat(pattern = "yyyy-MM-dd'T'H:mm:ss")
+  val eventTime: LocalDateTime? = null,
+  @Schema(description = "ID that references the event (optional)", required = false, example = "123456789")
+  val id: String? = null,
+  @Schema(description = "Json payload of data, normally no additional data would be sent", required = false)
+  val eventDetails: String? = null,
+)
