+++ src/main/kotlin/uk/gov/gdx/datashare/service/ConsumersService.kt
@@ -11,10 +11,8 @@ import java.util.*
 
 @Service
 @Transactional
-class SubscriptionManagerService(
-  private val eventSubscriptionRepository: EventSubscriptionRepository,
+class ConsumersService(
   private val consumerSubscriptionRepository: ConsumerSubscriptionRepository,
-  private val eventPublisherRepository: EventPublisherRepository,
   private val eventConsumerRepository: EventConsumerRepository,
   private val egressEventTypeRepository: EgressEventTypeRepository
 ) {
@@ -22,56 +20,22 @@ class SubscriptionManagerService(
     val log: Logger = LoggerFactory.getLogger(this::class.java)
   }
 
-  suspend fun getPublishers() = eventPublisherRepository.findAll()
-
-  suspend fun getEventSubscriptions() = eventSubscriptionRepository.findAll()
-
   suspend fun getConsumers() = eventConsumerRepository.findAll()
 
-  suspend fun getConsumersSubscriptions() = consumerSubscriptionRepository.findAll()
+  suspend fun getConsumerSubscriptions() = consumerSubscriptionRepository.findAll()
 
-  suspend fun addEventSubscription(
-    publisherSubRequest: PublisherSubRequest
-  ): EventSubscription {
-    with(publisherSubRequest) {
-      return eventSubscriptionRepository.save(
-        EventSubscription(
-          publisherId = publisherId,
-          clientId = clientId,
-          eventTypeId = eventTypeId,
-          datasetId = datasetId,
-          eventSubscriptionId = UUID.randomUUID()
-        )
-      )
-    }
-  }
-
-  suspend fun updateEventSubscription(
-    subscriptionId: UUID,
-    publisherSubRequest: PublisherSubRequest
-  ): EventSubscription {
-    with(publisherSubRequest) {
-      return eventSubscriptionRepository.save(
-        eventSubscriptionRepository.findById(subscriptionId)?.copy(
-          publisherId = publisherId,
-          clientId = clientId,
-          eventTypeId = eventTypeId,
-          datasetId = datasetId
-        ) ?: throw RuntimeException("Subscription not found")
-      )
-    }
-  }
+  suspend fun getSubscriptionsForConsumer(consumerId: UUID) = consumerSubscriptionRepository.findAllByConsumerId(consumerId)
 
   suspend fun addConsumerSubscription(
+    consumerId: UUID,
     consumerSubRequest: ConsumerSubRequest
   ): ConsumerSubscription {
     with(consumerSubRequest) {
       val consumer = eventConsumerRepository.findById(consumerId) ?: throw RuntimeException("Consumer $consumerId not found")
 
       val egressEventType = EgressEventType(
-        eventTypeId = UUID.randomUUID(),
-        ingressEventType = eventTypeId,
-        description = "$eventTypeId for ${consumer.consumerName}",
+        ingressEventType = ingressEventType,
+        description = "$ingressEventType for ${consumer.consumerName}",
         enrichmentFields = enrichmentFields
       )
 
@@ -92,17 +56,17 @@ class SubscriptionManagerService(
   }
 
   suspend fun updateConsumerSubscription(
-    subId: UUID,
+    consumerId: UUID,
+    subscriptionId: UUID,
     consumerSubRequest: ConsumerSubRequest
   ): ConsumerSubscription {
     with(consumerSubRequest) {
       val consumer = eventConsumerRepository.findById(consumerId) ?: throw RuntimeException("Consumer $consumerId not found")
 
-      val existingEgressEventType = egressEventTypeRepository.findByIngressEventTypeAndConsumerId(eventTypeId, consumerId)
+      val existingEgressEventType = egressEventTypeRepository.findByIngressEventTypeAndConsumerId(ingressEventType, consumerId)
       val egressEventType = existingEgressEventType ?: EgressEventType(
-        eventTypeId = UUID.randomUUID(),
-        ingressEventType = eventTypeId,
-        description = "$eventTypeId for ${consumer.consumerName}",
+        ingressEventType = ingressEventType,
+        description = "$ingressEventType for ${consumer.consumerName}",
         enrichmentFields = enrichmentFields
       )
       if (existingEgressEventType == null) {
@@ -110,7 +74,7 @@ class SubscriptionManagerService(
       }
 
       return consumerSubscriptionRepository.save(
-        consumerSubscriptionRepository.findById(subId)?.copy(
+        consumerSubscriptionRepository.findById(subscriptionId)?.copy(
           consumerId = consumerId,
           pollClientId = pollClientId,
           eventTypeId = egressEventType.eventTypeId,
@@ -118,18 +82,6 @@ class SubscriptionManagerService(
           pushUri = pushUri,
           ninoRequired = ninoRequired
         ) ?: throw RuntimeException("Subscription not found")
-      )
-    }
-  }
-
-  suspend fun addPublisher(
-    publisherRequest: PublisherRequest
-  ): EventPublisher {
-    with(publisherRequest) {
-      return eventPublisherRepository.save(
-        EventPublisher(
-          publisherName = name
-        )
       )
     }
   }
@@ -148,25 +100,10 @@ class SubscriptionManagerService(
 }
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@Schema(description = "Publisher Subscription Request")
-data class PublisherSubRequest(
-  @Schema(description = "Publisher ID", required = true, example = "00000000-0000-0001-0000-000000000000")
-  val publisherId: UUID,
-  @Schema(description = "Client ID", required = true, example = "a-client-id")
-  val clientId: String,
-  @Schema(description = "Events Type", required = true, example = "DEATH_NOTIFICATION")
-  val eventTypeId: String,
-  @Schema(description = "Data Set", required = true, example = "DEATH_LEN")
-  val datasetId: String
-)
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
 @Schema(description = "Consumer Subscription Request")
 data class ConsumerSubRequest(
-  @Schema(description = "Consumer ID", required = true, example = "00000000-0000-0001-0000-000000000000")
-  val consumerId: UUID,
   @Schema(description = "Events Type", required = true, example = "DEATH_NOTIFICATION")
-  val eventTypeId: String,
+  val ingressEventType: String,
   @Schema(description = "Client ID used to poll event platform", required = false, example = "a-polling-client")
   val pollClientId: String? = null,
   @Schema(description = "Client ID used to callback to event platform", required = false, example = "a-callback-client")
@@ -183,12 +120,5 @@ data class ConsumerSubRequest(
 @Schema(description = "Consumer Request")
 data class ConsumerRequest(
   @Schema(description = "Consumer name", required = true, example = "DWP")
-  val name: String,
-)
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@Schema(description = "Publisher Request")
-data class PublisherRequest(
-  @Schema(description = "Publisher name", required = true, example = "DWP")
   val name: String,
 )
